Added properties stroke-width and color. Default stroke width is too thin, but with a higher stroke-width value and combined with a transparent fill, you can make outline versions of the same icon.
Passing a color value will be used in fill="currentColor" now to create dual colored SVGs as svg backgrounds or as Content in pseudo elements. One can use Gulp Cheerio to target a specific "accent" ID labeled by a design program in each of the SVGs to append fill="currentColor" to the appropriate SVG path. Tested and working.